Update LICENSE
I saw some forks that seem to target 1.12.2 minecraft. Nothing wrong with this, but self-publishing our work is not allowed. 

If they have modifications to our version, I will create the curseforge project and set up profit share for them.

If they have a port to higher versions, It will be published on our CS page and they will receive fair profit share.